# KC_API/api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "KickConnect API",
    "description": "Auto-generated Swagger documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/get-event-list/{accountId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "description": "",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/get-active-event-list/{accountId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "description": "",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/get-event-by-id/{accountId}/{eventId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "description": "",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/update-event/{eventId}": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": "",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "eventName": {
                  "example": "any"
                },
                "existingEventName": {
                  "example": "any"
                },
                "existingEventDescription": {
                  "example": "any"
                },
                "eventDescription": {
                  "example": "any"
                },
                "isReservation": {
                  "example": "any"
                },
                "reservationCount": {
                  "example": "any"
                },
                "isCostToAttend": {
                  "example": "any"
                },
                "costToAttend": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/add-event/": {
      "post": {
        "tags": [
          "Event"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "accountId": {
                  "example": "any"
                },
                "eventName": {
                  "example": "any"
                },
                "eventDescription": {
                  "example": "any"
                },
                "isReservation": {
                  "example": "any"
                },
                "reservationCount": {
                  "example": "any"
                },
                "isCostToAttend": {
                  "example": "any"
                },
                "costToAttend": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/deactivate-event/{accountId}/{eventId}": {
      "delete": {
        "tags": [
          "Event"
        ],
        "description": "",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/get-accounts": {
      "get": {
        "tags": [
          "Account"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/add-account": {
      "post": {
        "tags": [
          "Account"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "accountName": {
                  "example": "any"
                },
                "accountPhone": {
                  "example": "any"
                },
                "accountEmail": {
                  "example": "any"
                },
                "accountAddress": {
                  "example": "any"
                },
                "accountCity": {
                  "example": "any"
                },
                "accountState": {
                  "example": "any"
                },
                "accountZip": {
                  "example": "any"
                },
                "isSuperAdmin": {
                  "example": "any"
                },
                "phone2": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}